split both miner and user into seperate sets of functionality
