+++ src/blockchain/network/participants/mod.rs
@@ -1,0 +1,7 @@
+//! Functionality for different network roles
+
+mod miners;
+mod shared;
+mod users;
+
+pub use self::{miners::miners_run, users::users_run};
